Jack

+++ officehours.js-e
@@ -1,10 +1,11 @@
+const ACK = "👍"
+const NAK = "🛑"
+
 const moment = require('moment')
 var x = 0
 
 var queue = []
 var dequeued = []
-
-var d = new Date();
 
 const OFFICE_HOURS = process.env.OFFICE_HOURS
 const TA_CHANNEL = process.env.TA_CHANNEL
@@ -31,6 +32,7 @@ const tas = {
         last_helped_time: 0
     },
 }
+
 function ready(message, index) {
 
     /**
@@ -38,20 +40,21 @@ function ready(message, index) {
      * This will be a permenant skip, and will not add it to the dequeue cache.
      */
 
+     if (index >= queue.length) return
+
     var msg = queue[index].message
     msg.reply(`${tas[message.author.id].name} is ready for you. Move to TA office.`)
     msg.delete()
     
     //Tells you time spent and people on queue.
     if ( tas[message.author.id].last_helped_time != 0) {
-    startTime = tas[message.author.id].last_helped_time
-    endTime = new Date();
-    var timeDiff = endTime - startTime; //in ms
-    timeDiff /= 1000;
-    var timespent = Math.round(timeDiff) / 60;
-    message.reply("You have spent " + timespent +  " minutes with that team. " + (queue.length - 1) +" people on the queue.");
-    }
-    else {
+        startTime = tas[message.author.id].last_helped_time
+        endTime = new Date();
+        var timeDiff = endTime - startTime; //in ms
+        timeDiff /= 1000;
+        var timespent = Math.round(timeDiff) / 60;
+        message.reply("You have spent " + timespent +  " minutes with that team. " + (queue.length - 1) +" people on the queue.");
+    } else {
         message.reply("Readying up. There are " + (queue.length - 1) +" people left on the queue.");
     }
     
@@ -59,15 +62,18 @@ function ready(message, index) {
     queue.splice(index, 1)
     tas[message.author.id].last_helped_time = new Date();
     
-    message.react("👍")
- 
+    message.react(ACK)
+}
+
+function index(member) {
+    for (var i = 0; i < queue.length; i++)
+        if (queue[i].member.id == member.id)
+            return i
+    return -1
 }
 
 function contains(member) {
-    for (var i = 0; i < queue.length; i++) 
-        if (queue[i].member.id == member.id) 
-            return true
-    return false
+    return index(member) != -1
 }
 
 /**
@@ -76,14 +82,14 @@ function contains(member) {
  * let them know on success
  */
 
-exports.onNext = (client, message, args) => {
+exports.onNext = (message, args) => {
     if (message.channel.id != OFFICE_HOURS) return // Behavior is only in the os-office-hours channel
     
     if (contains(message.author)) {
-        message.react("🛑")
+        message.react(NAK)
         message.reply("You are already on the queue.")
             .then(msg => {
-                msg.delete( { timeout: 5000 })
+                msg.delete({ timeout: 5000 })
                 message.delete({ timeout: 5000 })
             })
         return
@@ -96,7 +102,7 @@ exports.onNext = (client, message, args) => {
         timestamp: new Date()
     })
 
-    message.react("👍")
+    message.react(ACK)
 
     message.reply(`You are now #${queue.length} in the queue.`)
        .then(msg => {
@@ -113,38 +119,53 @@ exports.onNext = (client, message, args) => {
  * There is currently no Bot process for letting the user know it was an accident
  */
 
-exports.onUndo = (client, message) => {
+exports.onUndo = (message, args) => {
     if (TA_CHANNEL == message.channel.id) {
         if (dequeued.length == 0) {
-            message.react("🛑")
-            message.reply("```There is currently nothing in the dequeue cache.```")
+            message.react(NAK)
+            message.reply("```nimrod\nThere is currently nothing in the dequeue cache.```")
             return
         }
         queue.splice(0, 1, dequeued.pop())
-        message.reply("```All Done! Don't screw up next time!")
+        message.react(ACK)
+        message.reply("```nimrod\nDone! Don't screw up next time!```")
     }
 }
 
-exports.onQueue = (client, message) => {
+exports.onQueue = (message, args) => {
     if (TA_CHANNEL == message.channel.id) {
-        var body = ""
         if (queue.length == 0) {
-            body = "The queue is empty right now! :D"
-        } else {
-            for (var i = 0; i < queue.length; i++) {
-                var username = queue[i].member.username
-                var waitTime = moment(queue[i].timestamp).fromNow()
-                var desc = queue[i].desc
+            message.channel.send("```nimrod\nThe queue is currently empty```")
+            return
+        }
+        var body = ""
+        for (var i = 0; i < queue.length; i++) {
+            var username = queue[i].member.username
+            var waitTime = moment(queue[i].timestamp).fromNow()
+            var desc = queue[i].desc
 
-                body += `${i}) ${username} "${desc}"\t\t [${waitTime}]\n`
-            }
+            body += `${i}) ${username} "${desc}"\t\t [${waitTime}]\n`
         }
         message.channel.send("```nimrod\n" + body + "```")
-        return
     }
 }
 
-exports.onRemove = (client, message, args) => {
+// This potentially could be where the TA-leave functionality goes
+exports.onLeave = (message, args) => {
+    if (OFFICE_HOURS == message.channel.id) {
+        if (!contains(message.author)) {
+            message.react(NAK);
+            message.delete({ timeout: 10 * 1000 })
+            return
+        }
+
+        queue.splice(index(message.author), 1)
+        message.react(ACK)
+        message.delete({ timeout: 10 * 1000 })
+    }
+}
+
+exports.onRemove = (message, args) => {
     if (TA_CHANNEL != message.channel.id) return
     
     if (args.length == 0 || isNaN(args[0])) {
@@ -152,30 +173,57 @@ exports.onRemove = (client, message, args) => {
         message.reply("`!remove <index>`")
         return
     }
+
     var index = parseInt(args[0])
-    ready(message, index);
+    if (index >= queue.length) {
+        message.react(NAK)
+        message.reply("Invalid index.")
+        return
+    }
+
+    message.react(ACK)
+    queue.splice(index, 1)
 }
 
-exports.onReady = (client, message) => {
+exports.onReady = (message, args) => {
     if (TA_CHANNEL != message.channel.id) return
 
     if (queue.length == 0) {
-        message.reply("The queue is empty right now, crack open a beer")
+        message.channel.send("```nimrod\nThe queue is currently empty```")
         return
     }
-    //TO:DO calculate time spent with user.
-    ready(message, 0)
+
+    var index = 0
+    if (args.length > 0 && !isNaN(args[0]))
+        index = parseInt(args[0])
+    
+    if (index >= queue.length) {
+        message.react(NAK)
+        message.reply("Invalid index.")
+        return
+    }
+
+    ready(message, index)
 }
 
-exports.onOof = (client, message, args) => {
-     x++;
+exports.onOof = (message, args) => {
+     x++
      message.reply("There has been " + x + " 'persistent' questions to date.")
 }
 
-
-exports.onHelp = (client, message) => {
-    if (OFFICE_HOURS == message.channel.id)
-        message.reply("To join the queue, type ```next``` or ```!next``` followed by a brief description of what you need help with.")
-    else
-        message.reply("!queue to view the queue. !remove <index> to remove user, and notify them you're ready.")
+exports.onHelp = (message, args) => {
+    if (TA_CHANNEL == message.channel.id) {
+        message.reply("``` \
+            ping - simple test that responds with \"pong\". \
+            \n!queue - view the queue w/ username, issue description, and how long they've been waiting. \
+            \n!undo - quickly undo the ready command that removed them from the queue. \
+            \n!remove <index> - removes user from queue at certain index. Does not alert the user. \
+            \n!ready [index] - removes user from queue at index (top if index isn't provided). Alerts the user that the TA is ready. \
+            \n!help - shows these commands.```")
+        return
+    }
+    message.reply("``` \
+        next [issue] - adds a user to queue and responds with user's position in queue. Please provide an issue. \
+        \nhelp - provides a list of commands and their functions.```")
+        
 }
